コンポーネント分けた
